feat(comments): live @mention highlight, synced + scrollable panel, split files
Mentions
- Live highlight in the composer: a transparent textarea over a backdrop that
  paints the same text with the @name lit as a solid accent pill, so a tag is
  unmistakable the instant you type or pick someone (not just on the delivered
  comment). The pill is paint-only with net-zero padding so it never drifts off
  the caret. Delivered .mention strengthened to the same solid tag.

Panel scroll
- Wheel over the comments panel now scrolls the document (cards glide with their
  highlights), via a scroll-by message to the sandbox iframe. When a dense
  cluster of comments overflows the panel, the panel consumes the scroll first so
  the buried cards are reachable, then hands off to the document. Fixes comments
  piling up unreachable when many anchor to one spot and the doc cannot scroll.

Composer
- Close the composer optimistically on submit (capturing its anchor first): the
  send feels instant and the typed text never lingers beside the comment it became.

Refactor
- Split comment-thread.tsx (965 lines) into comment-composer.tsx (MentionField +
  Composer), comment-row.tsx (CommentBody + CommentRow), and comment-thread.tsx
  (PinnedZone + CommentCard + sections). Sort emoji.test imports.
